requirements.txt
# cognitive-swarm-agents/requirements.txt

# --- Configuration and Utilities ---
pydantic
pydantic-settings
python-dotenv

# --- Data Processing and PDF ---
arxiv
pymupdf # For PyMuPDF (PDF parsing)
pandas
numpy

# --- LLMs, Embeddings, Tokenization ---
openai # OpenAI Python client
tiktoken # OpenAI tokenizer

# --- LangChain Family ---
langchain-core
langchain-openai
langchain-community
langchain # Main LangChain package
langgraph

# --- LlamaIndex Family ---
llama-index-core
llama-index-vector-stores-mongodb
llama-index-embeddings-openai
llama-index-embeddings-huggingface # <<< NOUVELLE DÉPENDANCE AJOUTÉE ICI
llama-index-embeddings-ollama      # <<< AJOUT PRÉVENTIF POUR OLLAMA EMBEDDINGS LlamaIndex

# --- Database (MongoDB) ---
pymongo # Python driver for MongoDB
motor   # Async driver for MongoDB

# --- Evaluation and Metrics Logging ---
wandb # Weights & Biases

# --- Agent Frameworks ---
crewai 
# crewai-tools # Optional, often useful with crewai

# --- Web API (FastAPI) ---
fastapi
uvicorn[standard] # Includes standard dependencies for uvicorn like websockets, http-tools

# --- Development (Notebooks, etc.) ---
jupyter
ipykernel
# notebook # Optional, classic Jupyter Notebook interface

# Note: Specific versions should be pinned for production or stable environments.
# You can generate a more precise list from your active Conda environment (if pip was used within it):
# conda activate cognitive-swarm
# pip list --format=freeze > requirements.txt
# Or if you built a pure pip environment:
# source .venv/bin/activate (or .venv\Scripts\activate on Windows)
# pip freeze > requirements.txt